+++ accounts/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth import login as auth_login, authenticate, logout as aut
 from django.contrib.auth.decorators import login_required
 from .forms import CustomUserCreationForm, CustomErrorList
 from django.shortcuts import redirect
+from django.contrib.auth.models import User
 
 def signup(request):
     template_data = {}
@@ -38,3 +39,10 @@ def login(request):
 def logout(request):
     auth_logout(request)
     return redirect('home.index')
+
+@login_required
+def orders(request):
+    template_data = {}
+    template_data['title'] = 'Orders'
+    template_data['orders'] = request.user.order_set.all()
+    return render(request, 'accounts/orders.html', {'template_data': template_data})