+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Product models in the home page
 - Basic product reviews
 - Basic shopping cart
+- Order and Item models